Fix ROCm-LLVM EasyBlock after https://github.com/easybuilders/easybuild-easyblocks/pull/3799
With OMPT enabled, new issues were revealed.
We need to explicitly link our HSA library, as the internal (v1.0) is too old.
Also disable exporting the OpenMP headers (and other symbols), as this is done in
upstream LLVM, and causes issues when the symbols are copied (as not present where
expected).

Signed-off-by: Jan André Reuter <jan@zyten.de>